Fix Incremental... Issue #65835 -> #65843 #rb none #preflight 60820ac61046fb00019bcb9a
[CL 16100720 by David Lesage in ue5-main branch]

+++ Engine/Source/Runtime/Datasmith/CADKernel/Public/CADKernel/Geo/Sampling/PolylineTools.h
@@ -535,12 +535,12 @@ namespace CADKernel
 			OutCoordinates.Empty(EndIndex - StartIndex + 2);
 			OutPoints.Empty(EndIndex - StartIndex + 2);
 
-			if (FMath::IsNearlyEqual(PolylineCoordinates[StartIndex + 1], InBoundary.Min, KINDA_SMALL_NUMBER))
+			if (FMath::IsNearlyEqual(PolylineCoordinates[StartIndex + 1], InBoundary.Min, (double) KINDA_SMALL_NUMBER))
 			{
 				StartIndex++;
 			}
 
-			if (!FMath::IsNearlyEqual(PolylineCoordinates[EndIndex], InBoundary.Max, KINDA_SMALL_NUMBER))
+			if (!FMath::IsNearlyEqual(PolylineCoordinates[EndIndex], InBoundary.Max, (double) KINDA_SMALL_NUMBER))
 			{
 				EndIndex--;
 			}
